matcher/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
rapidfuzz==3.9.7
pandas==2.2.2
openai==1.51.0
pydantic==2.9.2
lxml==5.3.0
numpy==2.1.1
python-dotenv==1.0.1
